requirements.txt
flask
flask_session
python-dotenv
bcript
werkzeug
pdfkit